pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "f1flag"
version = "0.1.0"
description = "Monitor de banderas de F1 y NASCAR con control de bombilla inteligente"
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
dependencies = [
    "requests>=2.31.0",
    "python-kasa>=0.5.0",
    "websockets>=12.0",
    "python-dotenv>=1.0.0",
    "customtkinter>=5.2.0",
]

[tool.setuptools]
packages = ["racingflagmonitor"]

